Update
Code is fully functional. Insertion Angle check is commented out as it isn't necessary for project completion, but the rest of the checks work. Output is a list of all the checks for the lines, and the coordinates of the entry and target point for the best line. This is also saved as a separate markupfiducial as trajectory ready to be send to ROS via IGTLink